requirements.txt
gradio>=4.0.0
datasets>=2.14.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
huggingface_hub>=0.17.0
